requirements.txt
bleak
paho-mqtt
asyncio
